requirements.txt
aiohttp==3.6.2 
async-timeout==3.0.1   
azureml==0.2.7 
beautifulsoup4==4.8.2  
bs4==0.0.1 
chardet==3.0.4 
dj-database-url==0.5.0 
dj-static==0.0.6   
Django==3.0.5  
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-extra-fields==2.0.5
django-filter==2.2.0
djangorestframework==3.11.0
gpxpy==1.4.0
gunicorn==20.0.4
# lpip xml==3.7.2
Markdown==3.2.1
multidict==4.7.5   
oauthlib==3.0.1
olefile==0.46  
pandas==1.0.3  
Pillow==6.0.0  
psycopg2==2.8.4
python-dateutil==2.7.3 
python-decouple==3.3
pytz==2018.5   
requests==2.21.0   
requests-oauthlib==1.2.0   
six==1.13.0
static3==0.7.0 
telepot==12.7  
twitter==1.18.0
twython==3.8.2 
urllib3==1.24.2
virtualenv==20.0.18
whitenoise==5.0.1  
yarl==1.4.2
django-import-export
